updating file names to better organize project repo
