+++ frontend/src/pages/index.js
@@ -1,6 +1,8 @@
 // Re-export components from subdirectories
-export { default as BulletinBoard } from './bulletin_board';
-export { default as PersonalBoard } from './personal_board';
+export { default as PageBulletinBoard } from './page_bulletin_board';
+export { default as PagePersonalBoard } from './page_personal_board';
+export { default as PageAbout } from './page_about';
+export { default as HowTo } from './page_how_to';
 
 // Add more as you create them
 // export { default as About } from './about';
