+++ frontend/src/pages/page_about/about.css
