+++ frontend/src/pages/page_bulletin_board/bulletin_board.css
