+++ frontend/src/pages/page_bulletin_board/index.js
