+++ frontend/src/pages/page_how_to/how_to.css
