+++ frontend/src/pages/page_personal_board/personal_board.css
